Partial splint cleanup.

+++ libgpsd_core.c
@@ -812,7 +812,7 @@ char /*@observer@*/ *gpsd_id( /*@in@ */ struct gps_device_t *session)
 
 void gpsd_rollover_check(/*@in@ */ struct gps_device_t *session, const double unixtime)
 {
-    if (session->context->start_time >= GPS_EPOCH && unixtime < session->context->start_time)
+    if (session->context->start_time >= GPS_EPOCH && unixtime < (double)session->context->start_time)
 	gpsd_report(LOG_WARN, "GPS week rollover makes time invalid");
 }
